requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
pyyaml>=6.0
tqdm>=4.64.0 